+++ CHANGELOG.md
@@ -4,5 +4,8 @@
 * New API
 * New event onOSMapAddAdminSubmenu - allow to modify or add submenus
 * Removed J2.5 legacy code
+* Refined UI
+* Menu selector, click anywhere to select and better visual feedback
+
 
 ## 3.3.0 - Last release with J2.5 support